refactor: 컴포넌트 분리 및 비구조화할당 고려한 개선(step2)
